Add capability to list, create, read, update, and delete estimates

+++ src/Harvest/Expenses/Models/CreateExpense.cs
@@ -42,7 +42,7 @@ public class CreateExpense
     /// Gets or sets the quantity of units used to calculate the total cost of the expense.
     /// </summary>
     [JsonProperty("units")]
-    public int? Units { get; set; }
+    public decimal? Units { get; set; }
 
     /// <summary>
     /// Gets or sets the total cost of the expense.
